refactored all my code
