--- Animals/Buffalo/__init__.py
@@ -1,1 +1,0 @@
-from .request import Buffalo
