--- Animals/Catfish/request.py
@@ -1,10 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Catfish:
-
-    def __init__(self, name, species):
-        self.home = 'pond'
-        self.swimming = True
-
