--- Animals/Cobra/request.py
@@ -1,10 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Cobra:
-
-    def __init__(self, name, species):
-        self.home = 'glass tank'
-        self.slithering = True
-
