--- Animals/Crocodile/request.py
@@ -1,11 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Crocodile:
-
-    def __init__(self, name, species, food):
-        self.home = 'pond'
-        self.walking = True
-        self.swimming = True
-
