--- Animals/Goat/request.py
@@ -1,11 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Goat:
-
-    def __init__(self, name, species, shift):
-        self.home = 'petting area'
-        self.walking = True
-        self.shift = shift
-
