--- Animals/Iguana/request.py
@@ -1,11 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Iguana:
-
-    def __init__(self, name, species, shift):
-        self.home = 'glass tank'
-        self.walking = True
-        self.shift = shift
-
