--- Animals/King_snake/request.py
@@ -1,10 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class King_snake:
-
-    def __init__(self, name, species, food):
-        self.home = 'glass tank'
-        self.slithering = True
-
