--- Animals/Komodo_dragon/__init__.py
@@ -1,1 +1,0 @@
-from .request import Komodo_dragon
