--- Animals/Komodo_dragon/request.py
@@ -1,10 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Komodo_dragon:
-
-    def __init__(self, name, species, food):
-        self.home = 'glass tank'
-        self.walking = True
-
