--- Animals/Mallard/request.py
@@ -1,12 +1,0 @@
-from datetime import date
-from Animals.AnimalClass import Animal
-
-
-class Mallard:
-
-    def __init__(self, name, species, shift):
-        self.home = 'pond'
-        self.flying = True
-        self.swimming = True
-        self.walking = True
-
