+++ Animals/__init__.py
@@ -1,15 +1,13 @@
-from .Goat import Goat
-from .Bearded_dragon import Bearded_dragon
-from .Buffalo import Buffalo
-from .Camel import Camel
-from .Catfish import Catfish
-from .Cobra import Cobra
-from .Crocodile import Crocodile
-from .Donkey import Donkey
-from .Gecko import Gecko
-from .King_snake import King_snake
-from .Iguana import Iguana
-from .Komodo_dragon import Komodo_dragon
-from .Llama import Llama
-from .Mallard import Mallard
-from .Turtle import Turtle
+from .goat import Goat
+from .bearded_dragon import Bearded_dragon
+from .buffalo import Buffalo
+from .camel import Camel
+from .catfish import Catfish
+from .cobra import Cobra
+from .crocodile import Crocodile
+from .donkey import Donkey
+from .gecko import Gecko
+from .king_snake import King_snake
+from .iguana import Iguana
+from .llama import Llama
+from .turtle import Turtle
