+++ Animals/camel.py
@@ -1,0 +1,10 @@
+from datetime import date
+from .animal import Animal
+
+class Camel:
+
+    def __init__(self, name, species, shift):
+        self.home = 'petting area'
+        self.walking = True
+        self.shift = shift
+
