+++ Animals/catfish.py
@@ -1,0 +1,10 @@
+from datetime import date
+from .animal import Animal
+
+
+class Catfish:
+
+    def __init__(self, name, species):
+        self.home = 'pond'
+        self.swimming = True
+
