+++ Animals/crocodile.py
@@ -1,0 +1,11 @@
+from datetime import date
+from .animal import Animal
+
+
+class Crocodile:
+
+    def __init__(self, name, species, food):
+        self.home = 'pond'
+        self.walking = True
+        self.swimming = True
+
