+++ Animals/gecko.py
@@ -1,0 +1,10 @@
+from datetime import date
+from .animal import Animal
+
+
+class Gecko:
+
+    def __init__(self, name, species):
+        self.home = 'glass tank'
+        self.walking = True
+
