+++ Animals/iguana.py
@@ -1,0 +1,11 @@
+from datetime import date
+from .animal import Animal
+
+
+class Iguana:
+
+    def __init__(self, name, species, shift):
+        self.home = 'glass tank'
+        self.walking = True
+        self.shift = shift
+
