+++ Animals/llama.py
@@ -1,0 +1,11 @@
+from datetime import date
+from .animal import Animal
+
+
+class Llama:
+
+    def __init__(self, name, species, shift):
+        self.home = 'petting area'
+        self.walking = True
+        self.shift = shift
+
