--- Attractions/SnakePit/__init__.py
@@ -1,1 +1,0 @@
-from .request import SnakePit
