+++ Attractions/__init__.py
@@ -1,3 +1,3 @@
-from .PettingZoo import PettingZoo
-from .SnakePit import SnakePit
-from .Wetlands import Wetlands
+from .petting_zoo import PettingZoo
+from .snake_pit import SnakePit
+from .wetlands import Wetlands
